pyproject.toml
[tool.poetry]
name = "solarwinds_ipam"
version = "0.1.0"
description = "Solarwinds IP Address Management API Access"
authors = ["Jos Schilders <jschilders@groomlake.nl>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
rich = "^13.7.1"

[tool.poetry.dev-dependencies]
rich = "^13.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 118